+++ rope/base/oi/soi.py
@@ -4,6 +4,7 @@ For more information see the documentation in `rope.base.oi`
 package.
 
 """
+
 import rope.base.builtins  # Use full qualification for clarity.
 from rope.base import arguments, evaluate, pynames, pyobjects, utils
 from rope.base.oi.type_hinting.factory import get_type_hinting_factory