+++ rope/contrib/autoimport/sqlite.py
@@ -22,6 +22,7 @@ from rope.base.resources import Resource
 from rope.contrib.autoimport import models
 from rope.contrib.autoimport.defs import (
     ModuleFile,
+    Alias,
     Name,
     NameType,
     Package,
@@ -330,6 +331,13 @@ class AutoImport:
             yield SearchResult(
                 f"import {module}", module, source, NameType.Module.value
             )
+        for alias, module, source in self._execute(
+            models.Alias.search_modules_with_alias.select("alias", "module", "source"),
+            (name,),
+        ):
+            yield SearchResult(
+                f"import {module} as {alias}", alias, source, NameType.Module.value
+            )
 
     def get_modules(self, name) -> List[str]:
         """Get the list of modules that have global `name`."""
@@ -471,11 +479,14 @@ class AutoImport:
         """
         with self.connection:
             self._execute(models.Name.objects.drop_table())
+            self._execute(models.Alias.objects.drop_table())
             self._execute(models.Package.objects.drop_table())
             self._execute(models.Metadata.objects.drop_table())
             models.Name.create_table(self.connection)
+            models.Alias.create_table(self.connection)
             models.Package.create_table(self.connection)
             models.Metadata.create_table(self.connection)
+            self.add_aliases(self.project.prefs.autoimport.aliases)
             data = (
                 versioning.calculate_version_hash(self.project),
                 json.dumps(versioning.get_version_hash_data(self.project)),
@@ -594,6 +605,10 @@ class AutoImport:
             name.source.value,
             name.name_type.value,
         )
+
+    def add_aliases(self, aliases: Iterable[Alias]):
+        if aliases:
+            self._executemany(models.Alias.objects.insert_into(), aliases)
 
     def _add_names(self, names: Iterable[Name]):
         if names is not None: