+++ rope/contrib/finderrors.py
@@ -23,6 +23,7 @@ TODO:
 * ... ;-)
 
 """
+
 from rope.base import ast, evaluate, pyobjects